Additions to handle authorizer-information

+++ sam/src/lambda_delete_survey/index.js
@@ -1,36 +1,3 @@
-// const AWS = require('aws-sdk');
-// AWS.config.update({
-//     region: "eu-central-1"
-// });
-// // Create DynamoDB service object
-// var endPoint = (process.env.LOCAL_ENDPOINT == "AWS::NoValue") ? null : process.env.LOCAL_ENDPOINT;
-// var dynamoDb = new AWS.DynamoDB({ apiVersion: '2012-08-10', endpoint: endPoint });
-
-// exports.handler = async (event) => {
-//     // var id = parseInt(JSON.parse(event.body).id);
-//     var id = JSON.parse(event.body).id
-//     var params = {
-//         'TableName': `${process.env.TABLE_NAME}`,
-//         'Key': {
-//             'id':
-//             {
-//                 'S': `${id}`
-//             }
-//         }
-//     };
-//     var result;
-//     try {
-//         result = await dynamoDb.deleteItem(params).promise();
-//     } catch (err) {
-//         result = err;
-//     }
-//     const response = {
-//         statusCode: 200,
-//         body: JSON.stringify(result)
-//     };
-//     return response;
-// };
-
 cloudIntegration = require(process.env.AWS ? '/opt/aws-integration/index' : '../layers/aws-integration/index');
 
 exports.handler = async (event) => {
@@ -38,9 +5,25 @@ exports.handler = async (event) => {
 };
 
 async function businessLogic(event) {
-    var data = await cloudIntegration.SURVEY_REPOSITORY.deleteSurvey(event);
-    return {
-        executionSuccessful: true,
-        data
-    }
+    if (cloudIntegration.MODULE_PRIVILEGES_HELPER.isAdmin()) {
+        const uuidV4Regex = /^[A-F\d]{8}-[A-F\d]{4}-4[A-F\d]{3}-[89AB][A-F\d]{3}-[A-F\d]{12}$/i;
+        var surveyId = event['pathParameters']['surveyID'];
+        if (uuidV4Regex.test(surveyId)) {
+            var data = await cloudIntegration.SURVEY_REPOSITORY.deleteSurvey(event);
+        } else {
+            return {
+                executionSuccessful: false,
+                errorMessage: 'SurveyID invalid!'
+            }
+        }
+        return {
+            executionSuccessful: true,
+            data
+        }
+    } else {
+        return {
+            executionSuccessful: false,
+            errorMessage: 'No priviliges for requested action!'
+        }
+    } 
 }
